move TUF client into packages directory
