+++ packages/client/README.md
@@ -1,0 +1,3 @@
+# tuf-js
+
+JavaScript TUF client implementation.
