+++ packages/client/jest.config.js
@@ -1,0 +1,5 @@
+const base = require('../../jest.config.base')
+
+module.exports = {
+  ...base,
+};
