+++ packages/client/src/error.ts
@@ -10,9 +10,6 @@ export class PersistError extends Error {}
 // looking from the perspective of users of metadata API or ngclient.
 export class RepositoryError extends Error {}
 
-// An error about metadata object with insufficient threshold of signatures.
-export class UnsignedMetadataError extends RepositoryError {}
-
 // An error for metadata that contains an invalid version number.
 export class BadVersionError extends RepositoryError {}
 
@@ -21,13 +18,6 @@ export class EqualVersionError extends BadVersionError {}
 
 // Indicate that a TUF Metadata file has expired.
 export class ExpiredMetadataError extends RepositoryError {}
-
-// An error while checking the length and hash values of an object.
-export class LengthOrHashMismatchError extends RepositoryError {}
-
-export class CryptoError extends Error {}
-
-export class UnsupportedAlgorithmError extends CryptoError {}
 
 //----- Download Errors -------------------------------------------------------
