+++ packages/client/src/index.ts
@@ -1,3 +1,2 @@
 export { BaseFetcher } from './fetcher';
-export { TargetFile } from './models/file';
 export { Updater } from './updater';
