+++ packages/client/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist"
+  },
+  "exclude": ["./dist", "**/__tests__"],
+  "references": [
+    { "path": "../models" },
+    { "path": "../repo-mock" }
+  ]
+}
